fix: remove concept of consuming, renamed `BoundedConsumableBuffer` to `CircularBuffer`, cleaned up buffer API
